Move WindowPositionManager to own file. Update API
