+++ GitUI/GitExtensionsForm.cs
@@ -1,10 +1,8 @@
 using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
 using GitExtUtils.GitUI;
-using JetBrains.Annotations;
 using ResourceManager;
 
 namespace GitUI
@@ -48,7 +46,7 @@ namespace GitUI
                 }
 
                 needsPositionSave = false;
-                _windowPositionManager.SavePosition(this, GetType().Name);
+                _windowPositionManager.SavePosition(this);
                 TaskbarProgress.Clear();
             }
         }
@@ -96,7 +94,7 @@ namespace GitUI
 
             _windowCentred = StartPosition == FormStartPosition.CenterParent;
 
-            var position = _windowPositionManager.LookupWindowPosition(GetType().Name);
+            var position = _windowPositionManager.LoadPosition(this);
             if (position == null)
             {
                 return;
@@ -149,111 +147,6 @@ namespace GitUI
             }
 
             ResumeLayout();
-        }
-    }
-
-    internal sealed class WindowPositionManager
-    {
-        private static WindowPositionList _windowPositionList;
-
-        /// <summary>
-        ///   Restores the position of a form from the user settings. Does
-        ///   nothing if there is no entry for the form in the settings, or the
-        ///   setting would be invisible on the current display configuration.
-        /// </summary>
-        [Pure]
-        public Rectangle? FindWindowScreen(Point location, IEnumerable<Rectangle> desktopWorkingArea)
-        {
-            var distance = new SortedDictionary<float, Rectangle>();
-
-            foreach (var rect in desktopWorkingArea)
-            {
-                if (rect.Contains(location) && !distance.ContainsKey(0.0f))
-                {
-                    return null; // title in screen
-                }
-
-                int midPointX = rect.X + (rect.Width / 2);
-                int midPointY = rect.Y + (rect.Height / 2);
-                var d = (float)Math.Sqrt(((location.X - midPointX) * (location.X - midPointX)) +
-                                           ((location.Y - midPointY) * (location.Y - midPointY)));
-                distance.Add(d, rect);
-            }
-
-            return distance.FirstOrDefault().Value;
-        }
-
-        public WindowPosition LookupWindowPosition(string name)
-        {
-            try
-            {
-                if (_windowPositionList == null)
-                {
-                    _windowPositionList = WindowPositionList.Load();
-                }
-
-                var pos = _windowPositionList?.Get(name);
-
-                if (pos != null && !pos.Rect.IsEmpty)
-                {
-                    return pos;
-                }
-            }
-            catch
-            {
-                // TODO: how to restore a corrupted config?
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        ///   Save the position of a form to the user settings. Hides the window
-        ///   as a side-effect.
-        /// </summary>
-        /// <param name = "name">The name to use when writing the position to the
-        ///   settings</param>
-        public void SavePosition(Form form, string name)
-        {
-            try
-            {
-                var rectangle = form.WindowState == FormWindowState.Normal
-                    ? form.DesktopBounds
-                    : form.RestoreBounds;
-
-                var formWindowState = form.WindowState == FormWindowState.Maximized
-                    ? FormWindowState.Maximized
-                    : FormWindowState.Normal;
-
-                if (_windowPositionList == null)
-                {
-                    _windowPositionList = WindowPositionList.Load();
-                    if (_windowPositionList == null)
-                    {
-                        return;
-                    }
-                }
-
-                WindowPosition windowPosition = _windowPositionList.Get(name);
-                var windowCentred = form.StartPosition == FormStartPosition.CenterParent;
-
-                // Don't save location when we center modal form
-                if (windowPosition != null && form.Owner != null && windowCentred)
-                {
-                    if (rectangle.Width <= windowPosition.Rect.Width && rectangle.Height <= windowPosition.Rect.Height)
-                    {
-                        rectangle.Location = windowPosition.Rect.Location;
-                    }
-                }
-
-                var position = new WindowPosition(rectangle, DpiUtil.DpiX, formWindowState, name);
-                _windowPositionList.AddOrUpdate(position);
-                _windowPositionList.Save();
-            }
-            catch
-            {
-                // TODO: how to restore a corrupted config?
-            }
         }
     }
 }
