+++ GitUI/GitUI.csproj
@@ -1152,6 +1152,7 @@
     <Compile Include="WebBrowserEmulationMode.cs" />
     <Compile Include="WindowPositionList.cs" />
     <Compile Include="CommandsDialogs\CommitDialog\WordWrapper.cs" />
+    <Compile Include="WindowPositionManager.cs" />
     <Compile Include="WindowsJumpListManager.cs" />
     <Compile Include="WindowsThumbnailToolbarButton.cs" />
     <Compile Include="WindowsThumbnailToolbarButtons.cs" />
